Iia final/requirements.txt
mysql-connector-python
pandas
matplotlib
requests
numpy
Pillow